Modified initial states generator and few minor tweaks.

+++ ODE Abstraction/src/main/java/cz/muni/fi/ode/CoordinatePartitioner.java
@@ -10,5 +10,5 @@ import java.util.List;
  */
 public interface CoordinatePartitioner extends StateSpacePartitioner<CoordinateNode> {
 
-    public List<Range<Double>> getMyLimit();
+    public List<Range<Integer>> getMyLimit();
 }
